+++ R/print.R
@@ -24,7 +24,6 @@ knitr_is_rtf_output <- function() {
 
 knitr_is_word_output <- function() {
 
-  "word_document" %in% rmarkdown::all_output_formats(knitr::current_input()) |
   "docx" %in% knitr::opts_knit$get("rmarkdown.pandoc.to")
 }