+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Minor improvements and bug fixes
 
-* `docx` output format is now correctly dectected in **rmarkdown** and also in Quarto.
+* `docx` output format is now correctly detected in R Markdown and Quarto (#1040). (#1084)
 
 # gt 0.7.0