+++ NEWS.md
@@ -1,5 +1,9 @@
 # gt (development version)
 
+## Minor improvements and bug fixes
+
+* `docx` output format is now correctly detected in R Markdown and Quarto (#1040). (#1084)
+
 # gt 0.7.0
 
 ## New features